+++ include/gbui/widgets/containers.hpp
@@ -32,6 +32,7 @@
 #include "gbui/widgets/gallery.hpp"
 #include "gbui/widgets/marquee.hpp"
 #include "gbui/widgets/scroll.hpp"
+#include "gbui/widgets/splitPane.hpp"
 #include "gbui/widgets/table.hpp"
 #include "gbui/widgets/tabs.hpp"
 #include "gbui/widgets/treeView.hpp"